pulling LinearModel class up into _linear_impl
