+++ hummingbird/ml/operator_converters/skl_linear.py
@@ -7,49 +7,12 @@
 import torch
 import numpy as np
 from onnxconverter_common.registration import register_converter
+from ._linear_implementations import LinearModel
 
-from ._base_operator import BaseOperator
 
 """
 Converters for scikit-learn linear models: LinearRegression, LogisticRegression, LinearSVC, SGDClassifier, LogisticRegressionCV
 """
-
-
-class SklearnLinearModel(BaseOperator, torch.nn.Module):
-    def __init__(self, coefficients, intercepts, device, classes=[0], multi_class=None, is_linear_regression=False):
-        super(SklearnLinearModel, self).__init__()
-        self.coefficients = torch.nn.Parameter(torch.from_numpy(coefficients), requires_grad=False)
-        self.intercepts = torch.nn.Parameter(torch.from_numpy(intercepts), requires_grad=False)
-        self.classes = torch.nn.Parameter(torch.IntTensor(classes), requires_grad=False)
-        self.multi_class = multi_class
-        self.regression = is_linear_regression
-
-        self.perform_class_select = False
-        if min(classes) != 0 or max(classes) != len(classes) - 1:
-            self.perform_class_select = True
-
-        self.binary_classification = False
-        if len(classes) == 2:
-            self.binary_classification = True
-
-    def forward(self, x):
-        output = torch.addmm(self.intercepts, x, self.coefficients)
-        if self.multi_class == "multinomial":
-            output = torch.softmax(output, dim=1)
-        elif self.regression:
-            return output
-        else:
-            output = torch.sigmoid(output)
-            if not self.binary_classification:
-                output /= torch.sum(output, dim=1, keepdim=True)
-
-        if self.binary_classification:
-            output = torch.cat([1 - output, output], dim=1)
-
-        if self.perform_class_select:
-            return torch.index_select(self.classes, 0, torch.argmax(output, dim=1)), output
-        else:
-            return torch.argmax(output, dim=1), output
 
 
 def convert_sklearn_linear_model(operator, device, extra_config):
@@ -83,7 +46,7 @@ def convert_sklearn_linear_model(operator, device, extra_config):
         else:
             multi_class = "multinomial"
 
-    return SklearnLinearModel(coefficients, intercepts, device, classes=classes, multi_class=multi_class)
+    return LinearModel(coefficients, intercepts, device, classes=classes, multi_class=multi_class)
 
 
 def convert_sklearn_linear_regression_model(operator, device, extra_config):
@@ -102,7 +65,7 @@ def convert_sklearn_linear_regression_model(operator, device, extra_config):
     coefficients = operator.raw_operator.coef_.transpose().reshape(-1, 1).astype("float32")
     intercepts = operator.raw_operator.intercept_.reshape(1, -1).astype("float32")
 
-    return SklearnLinearModel(coefficients, intercepts, device, is_linear_regression=True)
+    return LinearModel(coefficients, intercepts, device, is_linear_regression=True)
 
 
 register_converter("SklearnLinearRegression", convert_sklearn_linear_regression_model)
